+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added `-A` / `--after-context` CLI flag to argument parser for specifying number of lines to show after each match.
+- Added docstrings to all test files.
+
+### Fixed
+
+- Resolved a series of pylint issues in the test files.
 
 ## [0.3.4] - 2025-10-21